ListView: fix default cell background colour
In Banshee's GTK2, the default cell background colour is white (taking
in account the default theme of Ubuntu 13.04). In Banshee's GTK3 however,
this was being drawn as grey, the same grey that is used for RulesHint
enabled rows and SortedColumn rows, so it was preventing the latter
effects to work. By using the "cell" StyleContext [1] instead of the
"view" one [2], this appears to be fixed.

[1] https://developer.gnome.org/gtk3/3.5/GtkStyleContext.html#GTK-STYLE-CLASS-CELL:CAPS
[2] https://developer.gnome.org/gtk3/3.5/GtkStyleContext.html#GTK-STYLE-CLASS-VIEW:CAPS