ListView: work-around possible selection GTK3 Theme CSS bug
In the case these two rules are defined in CSS:

*:selected {
    background-color: @selected_bg_color;
}

.cell {
    background-color: @base_color;
}

They have the same CSS precedence, so then the last one applies.
The solution to this is place the pseudo-class rule about the
:selected state the last one. This is bad because it is very common
to have this precedence clash in CSS files of Gtk themes, and
because it is actually the case in the default Ubuntu theme (at
least in 13.04), called Ambiance.

This bug has been reported upstream (I have even provided a patch):
https://bugs.launchpad.net/ubuntu/+source/ubuntu-themes/+bug/1211831

But in the meantime we can add this workaround in hyena, along with
a Hyena.Warning() to remind of the glitch.

Side note: This rendering glitch is also affecting the SourceView
treeview used in Banshee.ThickClient, and obviously this workaround
doesn't help there, so we will have to either find another temporary
workaround for that, or bug ubuntu-theme maintainers heavily.